Add sub-project for integration tests
 - Since there are overlaps with the integration tests and the unit tests running together (using redis binder vs test support binder), it is better to move the integration tests
to a separate sub-project
