+++ spring-cloud-task-integration-tests/src/test/resources/org/springframework/cloud/task/listener/sink-channel.properties
@@ -1,0 +1,1 @@
+spring.cloud.stream.bindings.input.destination=test
